Conflicts:
	app/code/Magento/Core/Model/AbstractModel.php
	app/code/Magento/Customer/Block/Account/Dashboard.php
	app/code/Magento/Customer/Block/Account/Dashboard/Newsletter.php
	app/code/Magento/Newsletter/Controller/Manage.php
	app/code/Magento/Newsletter/Controller/Subscriber.php
	app/code/Magento/Newsletter/Model/Observer.php
	app/code/Magento/Newsletter/Model/Resource/Problem/Collection.php
	app/code/Magento/Newsletter/Model/Resource/Subscriber/Collection.php
	app/code/Magento/Newsletter/Model/Subscriber.php
	app/code/Magento/Newsletter/etc/adminhtml/events.xml
	app/code/Magento/Newsletter/etc/frontend/events.xml
	app/code/Magento/Sales/Model/Quote.php
	dev/tests/integration/framework/Magento/TestFramework/TestCase/AbstractController.php
	dev/tests/integration/framework/tests/unit/testsuite/Magento/Test/TestCase/ControllerAbstractTest.php
	dev/tests/integration/testsuite/Magento/Backend/Utility/Controller.php
	dev/tests/integration/testsuite/Magento/Customer/Controller/AccountTest.php
	dev/tests/integration/testsuite/Magento/Customer/Controller/Adminhtml/IndexTest.php
	dev/tests/integration/testsuite/Magento/Newsletter/_files/subscribers.php
	dev/tests/static/testsuite/Magento/Test/Legacy/_files/obsolete_methods.php

+++ app/code/Magento/AdvancedCheckout/Model/Resource/Sku/Errors/Grid/Collection.php
@@ -64,9 +64,7 @@ class Collection extends \Magento\Data\Collection
             foreach ($parentBlock->getFailedItems() as $affectedItem) {
                 // Escape user-submitted input
                 if (isset($affectedItem['item']['qty'])) {
-                    $affectedItem['item']['qty'] = empty($affectedItem['item']['qty'])
-                        ? ''
-                        : (float)$affectedItem['item']['qty'];
+                    $affectedItem['item']['qty'] = empty($affectedItem['item']['qty']) ? '' : (double)$affectedItem['item']['qty'];
                 }
                 $item = new \Magento\Object();
                 $item->setCode($affectedItem['code']);
@@ -105,4 +103,3 @@ class Collection extends \Magento\Data\Collection
         return $this->_cart->getStore()->getWebsiteId();
     }
 }
-
